For 'ordering' column, TINYINT might be too small. Use SMALLINT instead.

+++ dp-devel/SETUP/upgrade/06/upgrade_news_table.php
@@ -40,7 +40,7 @@ dpsql_query("
         date_posted  INT         NOT NULL default 0,
         news_page_id VARCHAR(8)           default NULL,
         status       VARCHAR(8)  NOT NULL default '',
-        ordering     TINYINT     NOT NULL default '0',
+        ordering     SMALLINT    NOT NULL default '0',
         content      TEXT        NOT NULL,
         PRIMARY KEY id (id)
     ) TYPE=MyISAM
